Refactoring.., created sample
