+++ library/build.gradle
@@ -7,6 +7,7 @@ android {
     defaultConfig {
         minSdkVersion 8
         targetSdkVersion 22
+        resValue("string", "multiprocess_preferences_authority", "com.gdubina.multiprocesspreferences.PreferenceProvider")
     }
 
     buildTypes {
