+++ library/src/main/java/com/gdubina/multiprocesspreferences/Constants.java
@@ -1,0 +1,11 @@
+package com.gdubina.multiprocesspreferences;
+
+class Constants {
+    public static final String TYPE = "type";
+    public static final String KEY = "key";
+    public static final String INT_TYPE = "integer";
+    public static final String LONG_TYPE = "long";
+    public static final String FLOAT_TYPE = "float";
+    public static final String BOOLEAN_TYPE = "boolean";
+    public static final String STRING_TYPE = "string";
+}
