--- library/src/main/res/values/strings.xml
@@ -1,5 +1,0 @@
-<resources>
-
-    <string name="multiprocess_preferences_authority" translatable="false">com.gdubina.multiprocesspreferences.PREFFERENCE_AUTHORITY</string>
-
-</resources>
