+++ sample/src/androidTest/java/com/gdubina/multiprocesspreferences/sample/ApplicationTest.java
@@ -1,0 +1,15 @@
+package com.gdubina.multiprocesspreferences.sample;
+
+import android.app.Application;
+import android.test.ApplicationTestCase;
+
+/**
+ * <a href="http://d.android.com/tools/testing/testing_android.html">Testing Fundamentals</a>
+ */
+public class ApplicationTest extends ApplicationTestCase<Application>
+{
+    public ApplicationTest()
+    {
+        super(Application.class);
+    }
+}
