+++ settings.gradle
@@ -1,1 +1,1 @@
-include ':library'
+include ':library', ':sample'
